changed pressure to double

+++ app/controllers/StationCtrl.java
@@ -17,7 +17,7 @@ public class StationCtrl extends Controller
     render("station.html", station);
   }
 
-  public static void addReading(Long id, int code, double temperature, double windSpeed, int pressure, double windDirection)
+  public static void addReading(Long id, int code, double temperature, double windSpeed, double pressure, double windDirection)
   {
     Reading reading = new Reading(code, temperature, windSpeed, pressure, windDirection);
     Station station = Station.findById(id);
